backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic-settings==2.2.1
azure-ai-formrecognizer==3.3.3
openai==1.30.1
python-multipart==0.0.9
python-docx==1.1.2
aiofiles==23.2.1
tenacity==8.3.0
python-jose[cryptography]==3.3.0
aiohttp==3.9.5
httpx==0.27.2
